+++ main/manager-api/src/main/java/xiaozhi/modules/device/entity/DeviceAttributeEntity.java
@@ -25,11 +25,14 @@ public class DeviceAttributeEntity {
     @Schema(description = "设备ID（mac地址）")
     private String deviceId;
 
-    @Schema(description = "设备语言（en, zh-cn）")
+    @Schema(description = "设备规范语言码")
     private String language;
 
     @Schema(description = "最近检测到的蓝牙信标ID")
     private String lastBeaconId;
+
+    @Schema(description = "关联智能体名称")
+    private String agentName;
 
     @Schema(description = "创建者")
     @TableField(fill = FieldFill.INSERT)