+++ main/manager-api/src/main/java/xiaozhi/modules/security/config/ShiroConfig.java
@@ -93,6 +93,7 @@ public class ShiroConfig {
         filterMap.put("/config/**", "server");
         filterMap.put("/device/address-book/call", "server");
         filterMap.put("/device/event/report", "server");
+        filterMap.put("/device/rebind", "server");
         filterMap.put("/agent/chat-history/report", "server");
         filterMap.put("/agent/chat-history/download/**", "anon");
         filterMap.put("/agent/chat-summary/**", "server");